ATR-762: renamed code fix for PX1011 diagnostic, fixed namespaces in PX1009 and PX1011 tests
